requirements.txt
# ===============================================================
# Core ML/AI packages
# ===============================================================
torch>=2.0.0
numpy>=1.21.0,<2.0.0  # Numpy 2.0 호환성 이슈 방지
pandas>=1.3.0
scikit-learn>=1.0.0
lightgbm>=4.0.0
shap>=0.45.0          # TechnicalAgent XAI 기능용

# ===============================================================
# Data fetching and financial data
# ===============================================================
yfinance>=0.2.40      # Yahoo Finance API 클라이언트

# ===============================================================
# API and HTTP requests
# ===============================================================
requests>=2.25.0

# ===============================================================
# Environment variables
# ===============================================================
python-dotenv>=0.19.0

# ===============================================================
# Data serialization
# ===============================================================
joblib>=1.0.0

# ===============================================================
# Dashboard and visualization
# ===============================================================
streamlit>=1.28.0     # 대시보드
plotly>=5.0.0
matplotlib>=3.5.0     # 백테스트 차트 생성용

# ===============================================================
# LLM and AI packages
# ===============================================================
openai>=1.0.0         # GPT 모델 API
transformers>=4.30.0  # FinBERT 및 기타 트랜스포머 모델

# ===============================================================
# Date utilities
# ===============================================================
python-dateutil>=2.8.0  # relativedelta 사용

# ===============================================================
# Development (optional)
# ===============================================================
# pytest>=6.0.0         # 테스트 프레임워크 (필요시 주석 해제)
# jupyter>=1.0.0        # 주피터 노트북 (필요시 주석 해제)
# ipykernel>=6.0.0      # 주피터 노트북 실행용 (필요시 주석 해제)
